LayoutTests/platform/wpe/storage/domstorage/clear-expected.txt

Test basic dom storage .clear() functionality.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing sessionStorage
storage.clear()
PASS storage.length is 0
storage['FOO'] = 'MyFOO'
storage['BAR'] = 'MyBar'
PASS storage.length is 2
PASS storage['FOO'] is "MyFOO"
PASS storage['BAR'] is "MyBar"
storage.clear()
PASS storage.length is 0
PASS storage['FOO'] is undefined
PASS storage['BAR'] is undefined


Testing localStorage
storage.clear()
PASS storage.length is 0
storage['FOO'] = 'MyFOO'
FAIL storage['FOO'] = 'MyFOO' threw exception QuotaExceededError: The quota has been exceeded.
storage['BAR'] = 'MyBar'
FAIL storage['BAR'] = 'MyBar' threw exception QuotaExceededError: The quota has been exceeded.
FAIL storage.length should be 2. Was 0.
FAIL storage['FOO'] should be MyFOO (of type string). Was undefined (of type undefined).
FAIL storage['BAR'] should be MyBar (of type string). Was undefined (of type undefined).
storage.clear()
PASS storage.length is 0
PASS storage['FOO'] is undefined
PASS storage['BAR'] is undefined
PASS successfullyParsed is true

TEST COMPLETE
